requirements.txt
pandas>=2.2.0
numpy>=2.2.0
matplotlib>=3.10.0
scipy>=1.15.0
seaborn>=0.13.0
scikit-learn>=1.6.0
sentence-transformers>=4.0.0
torch>=2.7.0
pyyaml>=6.0.0
aiohttp>=3.12.0
groq>=0.27.0
dotenv>=0.9.0
tqdm>=0.4.0
diff-match-patch